+++ src/jig/diffconvert.py
@@ -20,6 +20,7 @@ from os.path import islink
 from difflib import SequenceMatcher
 
 from git.exc import BadObject
+from jig.conf import CODEC
 
 
 def describe_diff(a, b):
@@ -49,14 +50,14 @@ def describe_diff(a, b):
     for tag, i1, i2, j1, j2 in SequenceMatcher(None, a, b).get_opcodes():
         if tag == 'equal':
             for idx, line in enumerate(b[j1:j2]):
-                yield (idx + j1 + 1, ' ', unicode(line))
+                yield (idx + j1 + 1, ' ', line.decode(CODEC))
             continue
         if tag == 'replace' or tag == 'delete':
             for idx, line in enumerate(a[i1:i2]):
-                yield (idx + i1 + 1, '-', unicode(line))
+                yield (idx + i1 + 1, '-', line.decode(CODEC))
         if tag == 'replace' or tag == 'insert':
             for idx, line in enumerate(b[j1:j2]):
-                yield (idx + j1 + 1, '+', unicode(line))
+                yield (idx + j1 + 1, '+', line.decode(CODEC))
 
 
 class DiffType(object):